+++ lib/mime_builder.rb
@@ -1,6 +1,7 @@
 module MIMEBuilder
-  VERSION = '0.0.3'
+  VERSION = '0.0.4'
 
   autoload :Filepath, 'mime_builder/file'
+  autoload :JSON, 'mime_builder/json'
   autoload :Text, 'mime_builder/text'
 end